ai_index_number/requirements.txt
streamlit
pandas
PyMuPDF
sentence-transformers
faiss-cpu
numpy
scikit-learn
rank-bm25
python-dotenv
ollama
pytest
